+++ packages/react-server/src/client.tsx
@@ -1,4 +1,4 @@
 "use client";
 
 export { Link } from "./lib/components/link";
-export { __history } from "./lib/csr";
+export { ErrorBoundary } from "./lib/components/error-boundary";
