+++ packages/react-server/src/index.ts
@@ -1,1 +1,1 @@
-export { createError, getErrorContext as getErrorStatus } from "./lib/error";
+export {};
