+++ packages/react-server/src/server.ts
@@ -1,5 +1,7 @@
+// TODO: export type from root?
 export type {
   PageRouteProps,
   LayoutRouteProps,
   ErrorRouteProps,
 } from "./lib/router";
+export { createError } from "./lib/error";
